added some comments

+++ geom.c
@@ -97,8 +97,8 @@ void find_collinear_straightforward(point2D* p, int n) {
         for (int j = i + 1; j < n - 1; j++) {
             point2D p2 = p[j];
 
-            double slope = calc_slope((double)p1.x, (double)p1.y, (double)p2.x, (double)p2.y);
-            double yIntercept = calc_y_intercept((double)p1.x, (double)p1.y, slope);
+            double slope = calc_slope(p1.x, p1.y, p2.x, p2.y);
+            double yIntercept = calc_y_intercept(p1.x, p1.y, slope);
             
             if (slope < DBL_MAX) {
                 // If the slope is not vertical
@@ -106,7 +106,7 @@ void find_collinear_straightforward(point2D* p, int n) {
                     point2D p3 = p[k];
                     
                     //Check to see if third point is colinear with line segement of first two points
-                    if (is_colinear((double)p3.x, (double)p3.y, slope, yIntercept)) {
+                    if (is_colinear(p3.x, p3.y, slope, yIntercept)) {
                         ncol++;
 //                        printf("Points: %i, %i, %i\n", i, j, k);
                     }
@@ -115,7 +115,7 @@ void find_collinear_straightforward(point2D* p, int n) {
                 // Vertical slope
                 for (int k = j + 1; k < n; k++) {
                     // Check if the third point has a shared x coordinate
-                    if ((double)p[k].x == (double)p1.x) {
+                    if (p[k].x == p1.x) {
                         ncol++;
 //                        printf("Vertical line segment. Points: %i, %i, %i \n", i, j, k);
                     }
@@ -158,7 +158,7 @@ void find_collinear_improved(point2D* p, int n) {
             
             //Create struct lnSegment with two points (i and j) and current slope
             struct LnSegment ln;
-            ln.slope = calc_slope((double)p1.x, (double)p1.y, (double)p2.x, (double)p2.y);
+            ln.slope = calc_slope(p1.x, p1.y, p2.x, p2.y);
             ln.x1 = p1.x;
             ln.y1 = p1.y;
             ln.x2 = p2.x;
@@ -207,6 +207,7 @@ void find_collinear_improved(point2D* p, int n) {
         struct LnSegment ln1;
         struct LnSegment ln2;
         for (int a = 0; a < size - 1; a++) {
+            // Number of slopes with the same value
             int count = 1;
             ln1 = lines[a];
             ln2 = lines[a + 1];
@@ -215,6 +216,7 @@ void find_collinear_improved(point2D* p, int n) {
                 if (a + count >= size) {
                     break;
                 }
+                // Check next line
                 ln2 = lines[a + count];
             }
             if (count > 1) {
@@ -225,6 +227,7 @@ void find_collinear_improved(point2D* p, int n) {
                 int numOfTriplets = factorial(count + 1) / (factorial(3) * factorial(count + 1 - 3));
                 ncol = ncol + numOfTriplets;
             }
+            // Move index a along
             a = a + count - 1;
         }
     }
